Rename utils.mock_vasp module to utils.mock_code
